feat: Expose register_contract_mutations from essential-node-types
